Conflicts:
	Phat Beats/resource/Sound/music credits.txt
	Phat Beats/source/CPlayer.cpp
	Phat Beats/source/States/CGameplay_State.cpp
	Phat Beats/source/States/CMenu_State.cpp

+++ Phat Beats/source/States/CGameplay_State.h
@@ -64,7 +64,7 @@ private:
 	//*******MEMBERS**********//
 	Bitmap_Font* m_bMenu_Font;
 	bool m_bPreviouslyPlaying;
-	int m_SongTransitionAlpha;
+	float m_SongTransitionAlpha;
 	bool m_bStartTransition;
 	// Managers
 	CBeatManager* BeatManager;
